Add taxonomic information to a Data Frame obtained from HMMER with a "hits.taxid" column.

Add taxonomic information to a Data Frame obtained from HMMER with a "hits.taxid" column.

add_taxa_to_hmmer_tbl(data, mode = "remote", rank_vc = NULL)

Arguments

data

An Data Frame obtained from search function.

mode

Either "local" or "remote". If "local" you will use a local database instead of remote resources. You will not have to download the database but it is slower.

rank_vc

A character vector containing the desired taxonomic ranks. If empty, all available taxonomic ranges will be retrieved.

Value

A Data Frame with new taxonomic parameters.

Examples

data(phmmer_2abl)
add_taxa_to_hmmer_tbl(
    data = phmmer_2abl,
    mode = "remote",
    rank_vc = NULL
)
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> No ENTREZ API key provided
#>  Get one via taxize::use_entrez()
#> See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
#> # A tibble: 25 × 92
#>    taxa.taxid taxa.no …¹ taxa.…² taxa.…³ taxa.…⁴ taxa.…⁵ taxa.…⁶ taxa.…⁷ taxa.…⁸
#>    <chr>      <chr>      <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 10090      cellular … Eukary… Opisth… Metazoa Eumeta… Bilate… Deuter… Chorda…
#>  2 10090      cellular … Eukary… Opisth… Metazoa Eumeta… Bilate… Deuter… Chorda…
#>  3 10090      cellular … Eukary… Opisth… Metazoa Eumeta… Bilate… Deuter… Chorda…
#>  4 10090      cellular … Eukary… Opisth… Metazoa Eumeta… Bilate… Deuter… Chorda…
#>  5 10116      cellular … Eukary… Opisth… Metazoa Eumeta… Bilate… Deuter… Chorda…
#>  6 10116      cellular … Eukary… Opisth… Metazoa Eumeta… Bilate… Deuter… Chorda…
#>  7 10116      cellular … Eukary… Opisth… Metazoa Eumeta… Bilate… Deuter… Chorda…
#>  8 11776      Feline sa… Viruses Ribovi… Pararn… NA      NA      NA      Artver…
#>  9 237561     cellular … Eukary… Opisth… Fungi   saccha… CUG-Se… Candid… Ascomy…
#> 10 237631     cellular … Eukary… Opisth… Fungi   NA      NA      NA      Basidi…
#> # … with 15 more rows, 83 more variables: taxa.subphylum <chr>,
#> #   taxa.clade.4 <chr>, taxa.clade.5 <chr>, taxa.clade.6 <chr>,
#> #   taxa.clade.7 <chr>, taxa.superclass <chr>, taxa.clade.8 <chr>,
#> #   taxa.clade.9 <chr>, taxa.clade.10 <chr>, taxa.class <chr>,
#> #   taxa.clade.11 <chr>, taxa.clade.12 <chr>, taxa.clade.13 <chr>,
#> #   taxa.superorder <chr>, taxa.clade.14 <chr>, taxa.order <chr>,
#> #   taxa.suborder <chr>, taxa.clade.15 <chr>, taxa.family <chr>, …