hmmscan_2abl

A DataFrame with 4 results of a hmmscan search of the 2abl sequence.

hmmscan_2abl

Format

A data frame with 36 variables:

algorithm

HMMER algorithm

uuid

unique hmmer identifier

stats.page

stats.nhits

The number of hits found above reporting thresholds

stats.elapsed

stats.Z

The number of sequences or models in the target database

stats.Z_setby

stats.n_past_msv

stats.unpacked

stats.nseqs

stats.user

stats.domZ_setby

stats.n_past_bias

stats.sys

stats.n_past_fwd

stats.total

stats.nmodels

The number of models in this search

stats.nincluded

The number of sequences or models scoring above the significance threshold

stats.n_past_vit

stats.nreported

The number of sequences or models scoring above the reporting threshold

stats.domZ

The number of hits in the target database

hits.acc

Accession of the target

hits.desc

Description of the target

hits.pvalue

P-value of the score

hits.flags

hits.nregions

hits.taxid

hits.nreported

hits.hindex

hits.ndom

Total number of domains identified in this sequence

hits.name

Name of the target (sequence for phmmer/hmmsearch, HMM for hmmscan)

hits.score

Bit score of the sequence (all domains, without correction)

hits.bias

hits.domains

The domain or hit hash contains the details of the match, in particular the alignment between the query and the target.

hits.evalue

E-value of the score

hits.nincluded

E-value of the score

For further details, see https://hmmer-web-docs.readthedocs.io/en/latest/appendices.html