Perform a hmmsearch search of a protein alignment against a protein sequence database.

Perform a hmmsearch search of a protein alignment against a protein sequence database.

search_hmmsearch(
  alns,
  aln_names = NULL,
  dbs = "swissprot",
  verbose = TRUE,
  N.TRIES = 1
)

Arguments

alns

A Biostrings::AAMultipleAlignment or a list of Biostrings::AAMultipleAlignment.

aln_names

A character vector containing the names of the aligments. If alns is a named list it will overwrite them. If not specified, they will not be taken into account.

dbs

A character vector containing the target databases. Frequently used databases are swissprot, uniprotrefprot, uniprotkb, ensembl, pdb and alphafold, but a complete and updated list is available at https://www.ebi.ac.uk/Tools/hmmer/.

verbose

A logical, if TRUE details of the download process is printed.

N.TRIES

An integer specifying the number of trials before a time out occurs.

Value

An AnnotatedDataFrame, consisting of 2 parts, a nested DataFrame with the search hashes, the download links of all available files and of the HMMER page where the results are hosted, and the metadata associated to this DataFrame. Although all available results are available here, we recommend using the extract_from_HMMER_data_tbl function to preprocess the data.

Examples

path_to_example_aln <- system.file(
  "extdata/alignment_example.afa",
  package = "HMMERutils")
alns <- Biostrings::readAAMultipleAlignment(path_to_example_aln)
hmmsearch_tbl <- search_hmmsearch(
  alns = alns,
  dbs = "pdb",
  verbose = FALSE)
hmmsearch_tbl
#> An object of class 'AnnotatedDataFrame'
#>   HMMERqueryNames: 1
#>   varLabels: dbs uuid ... phylip.url (17 total)
#>   varMetadata: labelDescription algorithm
Biobase::varMetadata(hmmsearch_tbl)
#>                                                                                  labelDescription
#> dbs                                                  Names of target databases used in the search
#> uuid                                                                    The unique job identifier
#> score.url                                                            URL with HMMER results score
#> stats                                                                              The stats hash
#> hits                                                                     Array of sequence hashes
#> domains                                                          Array of sequence domains hashes
#> text.url                              A plain text file containing the hit alignments and scores.
#> tab.url                  A tab delimited text file containing the hit information. No alignments.
#> xml.url                                    An XML file formatted for machine parsing of the data.
#> json.url                             All the results information encoded as a single JSON string.
#> fasta.url                 Download the significant hits from your search as a gzipped FASTA file.
#> fullfasta.url                                                                   Full length FASTA
#> alignedfasta.url A gzipped file containing the full length sequences for significant search hits.
#> stockholm.url          A gzipped file containing aligned significant search hits in FASTA format.
#> clustalw.url                Download an alignment of significant hits as a gzipped ClustalW file.
#> psiblast.url                Download an alignment of significant hits as a gzipped psiblast file.
#> phylip.url                    Download an alignment of significant hits as a gzipped phylip file.
#>                  algorithm
#> dbs              hmmsearch
#> uuid             hmmsearch
#> score.url        hmmsearch
#> stats            hmmsearch
#> hits             hmmsearch
#> domains          hmmsearch
#> text.url         hmmsearch
#> tab.url          hmmsearch
#> xml.url          hmmsearch
#> json.url         hmmsearch
#> fasta.url        hmmsearch
#> fullfasta.url    hmmsearch
#> alignedfasta.url hmmsearch
#> stockholm.url    hmmsearch
#> clustalw.url     hmmsearch
#> psiblast.url     hmmsearch
#> phylip.url       hmmsearch