Perform an iterative search of a protein sequence or an alignment against a protein sequence database.

Perform an iterative search of a protein sequence or an alignment against a protein sequence database.

search_jackhmmer(
  seqs = NULL,
  alns = NULL,
  seq_or_aln_names = NULL,
  dbs = "swissprot",
  verbose = TRUE,
  N.TRIES = 1
)

Arguments

seqs

A character vector containing the sequences of the query or any other object that can be converted to that.

alns

A Biostrings::AAMultipleAlignment or a list of Biostrings::AAMultipleAlignment.

seq_or_aln_names

A character vector containing the names of the sequences or the alignments. If alns or seqs are named,it will overwrite them. If not specified, they will not be taken into account.

dbs

A character vector containing the target databases. Frequently used databases are swissprot, uniprotrefprot, uniprotkb, ensembl, pdb and alphafold, but a complete and updated list is available at https://www.ebi.ac.uk/Tools/hmmer/.

verbose

A logical, if TRUE details of the download process is printed.

N.TRIES

An integer specifying the number of attempts before an error occurs.

Value

An AnnotatedDataFrame, consisting of 2 parts, a nested DataFrame with the search hashes, the download links of all available files and of the HMMER page where the results are hosted, and the metadata associated to this DataFrame. Although all available results are available here, we recommend using the extract_from_HMMER_data_tbl function to preprocess the data.

Examples

jackhmmer_tbl_using_seq <- search_jackhmmer(
   seqs = "MTEITAAMVKELRTGAGMMDCKN",
  seq_or_aln_names = "1efu_B",
  dbs = "pdb",
  verbose = FALSE)
jackhmmer_tbl_using_seq
#> An object of class 'AnnotatedDataFrame'
#>   HMMERqueryNames: 1
#>   varLabels: seq.name dbs ... phylip.url (18 total)
#>   varMetadata: labelDescription algorithm
Biobase::varMetadata(jackhmmer_tbl_using_seq)
#>                                                                                  labelDescription
#> seq.name                                                    Names of sequences used in the search
#> dbs                                                  Names of target databases used in the search
#> uuid                                                                    The unique job identifier
#> score.url                                                            URL with HMMER results score
#> stats                                                                              The stats hash
#> hits                                                                     Array of sequence hashes
#> domains                                                          Array of sequence domains hashes
#> text.url                              A plain text file containing the hit alignments and scores.
#> tab.url                  A tab delimited text file containing the hit information. No alignments.
#> xml.url                                    An XML file formatted for machine parsing of the data.
#> json.url                             All the results information encoded as a single JSON string.
#> fasta.url                 Download the significant hits from your search as a gzipped FASTA file.
#> fullfasta.url                                                                   Full length FASTA
#> alignedfasta.url A gzipped file containing the full length sequences for significant search hits.
#> stockholm.url          A gzipped file containing aligned significant search hits in FASTA format.
#> clustalw.url                Download an alignment of significant hits as a gzipped ClustalW file.
#> psiblast.url                Download an alignment of significant hits as a gzipped psiblast file.
#> phylip.url                    Download an alignment of significant hits as a gzipped phylip file.
#>                  algorithm
#> seq.name         jackhmmer
#> dbs              jackhmmer
#> uuid             jackhmmer
#> score.url        jackhmmer
#> stats            jackhmmer
#> hits             jackhmmer
#> domains          jackhmmer
#> text.url         jackhmmer
#> tab.url          jackhmmer
#> xml.url          jackhmmer
#> json.url         jackhmmer
#> fasta.url        jackhmmer
#> fullfasta.url    jackhmmer
#> alignedfasta.url jackhmmer
#> stockholm.url    jackhmmer
#> clustalw.url     jackhmmer
#> psiblast.url     jackhmmer
#> phylip.url       jackhmmer

path_to_example_aln <- system.file(
  "extdata/alignment_example.afa",
  package = "HMMERutils")
alns <- Biostrings::readAAMultipleAlignment(path_to_example_aln)
jackhmmer_tbl_using_aln <- search_jackhmmer(
  alns = alns,
  dbs = "pdb",
  verbose = TRUE)
#> Content from HMMER server:
#> <opt>
#>   <data name="results" algo="hmmsearch" uuid="B79E9604-D1E5-11EC-9A70-2D1AE976C163">
#>     <_internal highevalue="3.0e-09" lowevalue="1.1e-08" />
#>     <hits name="3mmp_A" acc="3mmp_A" arch="PF00889.22 PF00009.30 PF03144.28 PF03143.20" archScore="5" archindex="118498722050600 83333" bias="6.0" desc="Elongation factor Tu 2, Elongation factor Ts" evalue="3.0e-09" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/3mmp_A" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="2" ph="Proteobacteria" pvalue="-32.8472391366523" score="41.1" sindex="57514389" species="Escherichia coli K-12" taxid="83333" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.863636363636364" aliIdCount="19" aliL="678" aliM="23" aliN="24" aliSim="0.863636363636364" aliSimCount="19" aliaseq="MAEITASLVKELRerTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="1" alihmmname="Query" alihmmto="22" alimline="m EiTA  VKELR  TGAGMMDCK" alimmline="" alimodel="mtEiTAAMVKELR..TGAGMMDCK" aliniseqs="2" alintseq="" alippline="99**********555********9" alirfline="" alisindex="57514389" alisqacc="3mmp_A" alisqdesc="Elongation factor Tu 2, Elongation factor Ts" alisqfrom="1" alisqname="3mmp_A" alisqto="24" bias="2.51" bitscore="39.9151382446289" cevalue="1.2e-13" iali="1" ienv="1" ievalue="7.2e-09" is_included="1" is_reported="1" jali="24" jenv="25" oasc="0.91" />
#>       <domains aliId="0.583333333333333" aliIdCount="7" aliL="678" aliM="23" aliN="12" aliSim="0.666666666666667" aliSimCount="8" aliaseq="VKNMITGAAQMD" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="20" alimline="VK + TGA  MD" alimmline="" alimodel="VKELRTGAGMMD" aliniseqs="2" alintseq="" alippline="99999***9998" alirfline="" alisindex="57514389" alisqacc="3mmp_A" alisqdesc="Elongation factor Tu 2, Elongation factor Ts" alisqfrom="373" alisqname="3mmp_A" alisqto="384" bias="0.07" bitscore="1.16900599002838" cevalue="0.18" iali="373" ienv="370" ievalue="11000" is_included="0" is_reported="1" jali="384" jenv="386" oasc="0.86" />
#>       <pdbs>3mmp_C</pdbs>
#>       <pdbs>3mmp_A</pdbs>
#>       <seqs an="" de="Elongation factor Tu 2, Elongation factor Ts" dn="3mmp_C" sp="Escherichia coli K-12" />
#>     </hits>
#>     <hits name="4r71_A" acc="4r71_A" arch="PF00889.22 PF00009.30 PF03144.28 PF03143.20" archScore="5" archindex="118498722050600 83333" bias="6.0" desc="Elongation factor Ts, Elongation factor Tu" evalue="3.1e-09" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/4r71_A" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="2" ph="Proteobacteria" pvalue="-32.7913793508402" score="41.1" sindex="57510549" species="Escherichia coli K-12" taxid="83333" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.863636363636364" aliIdCount="19" aliL="694" aliM="23" aliN="24" aliSim="0.863636363636364" aliSimCount="19" aliaseq="MAEITASLVKELRerTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="1" alihmmname="Query" alihmmto="22" alimline="m EiTA  VKELR  TGAGMMDCK" alimmline="" alimodel="mtEiTAAMVKELR..TGAGMMDCK" aliniseqs="2" alintseq="" alippline="99**********555********9" alirfline="" alisindex="57510549" alisqacc="4r71_A" alisqdesc="Elongation factor Ts, Elongation factor Tu" alisqfrom="1" alisqname="4r71_A" alisqto="24" bias="2.51" bitscore="39.8780784606934" cevalue="1.2e-13" iali="1" ienv="1" ievalue="7.4e-09" is_included="1" is_reported="1" jali="24" jenv="25" oasc="0.91" />
#>       <domains aliId="0.583333333333333" aliIdCount="7" aliL="694" aliM="23" aliN="12" aliSim="0.666666666666667" aliSimCount="8" aliaseq="VKNMITGAAQMD" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="20" alimline="VK + TGA  MD" alimmline="" alimodel="VKELRTGAGMMD" aliniseqs="2" alintseq="" alippline="99999***9998" alirfline="" alisindex="57510549" alisqacc="4r71_A" alisqdesc="Elongation factor Ts, Elongation factor Tu" alisqfrom="373" alisqname="4r71_A" alisqto="384" bias="0.07" bitscore="1.13319849967957" cevalue="0.19" iali="373" ienv="370" ievalue="11000" is_included="0" is_reported="1" jali="384" jenv="386" oasc="0.86" />
#>       <pdbs>4r71_C</pdbs>
#>       <pdbs>4r71_A</pdbs>
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu" dn="4r71_C" sp="Escherichia coli K-12" />
#>     </hits>
#>     <hits name="4pc6_C" acc="4pc6_C" arch="PF00889.22" archScore="2" archindex="158710124293725 536056" bias="1.1" desc="Elongation factor Ts" evalue="7.0e-09" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/4pc6_C" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="1" ph="Proteobacteria" pvalue="-31.9955245624493" score="40.0" sindex="55721010" species="Escherichia coli (strain ATCC 33849 / DSM 4235 / NCIMB 12045 / K12 / DH1)" taxid="536056" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.9" aliIdCount="18" aliL="282" aliM="23" aliN="22" aliSim="0.9" aliSimCount="18" aliaseq="EITASLVKELRerTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="3" alihmmname="Query" alihmmto="22" alimline="EiTA  VKELR  TGAGMMDCK" alimmline="" alimodel="EiTAAMVKELR..TGAGMMDCK" aliniseqs="2" alintseq="" alippline="9*********555********9" alirfline="" alisindex="55721010" alisqacc="4pc6_C" alisqdesc="Elongation factor Ts" alisqfrom="2" alisqname="4pc6_C" alisqto="23" bias="1.14" bitscore="39.962963104248" cevalue="1.1e-13" iali="2" ienv="1" ievalue="7.0e-09" is_included="1" is_reported="1" jali="23" jenv="24" oasc="0.90" />
#>       <domains aliId="0.8" aliIdCount="4" aliL="282" aliM="23" aliN="5" aliSim="0.8" aliSimCount="4" aliaseq="AAMSK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="6" alihmmname="Query" alihmmto="10" alimline="AAM K" alimmline="" alimodel="AAMVK" aliniseqs="2" alintseq="" alippline="9**77" alirfline="" alisindex="55721010" alisqacc="4pc6_C" alisqdesc="Elongation factor Ts" alisqfrom="276" alisqname="4pc6_C" alisqto="280" bias="0.41" bitscore="-3.55217576026917" cevalue="5.6" iali="276" ienv="275" ievalue="340000" is_included="0" is_reported="0" jali="280" jenv="280" oasc="0.82" />
#>       <pdbs>4pc6_D</pdbs>
#>       <pdbs>4pc6_C</pdbs>
#>       <seqs an="" de="Elongation factor Ts" dn="4pc6_D" sp="Escherichia coli DH1" />
#>     </hits>
#>     <hits name="4pc3_C" acc="4pc3_C" arch="PF00889.22" archScore="2" archindex="158710124293725 83333" bias="1.1" desc="Elongation factor Ts" evalue="7.0e-09" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/4pc3_C" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="1" ph="Proteobacteria" pvalue="-31.9955245624493" score="40.0" sindex="57516019" species="Escherichia coli K-12" taxid="83333" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.9" aliIdCount="18" aliL="282" aliM="23" aliN="22" aliSim="0.9" aliSimCount="18" aliaseq="EITASLVKELRerTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="3" alihmmname="Query" alihmmto="22" alimline="EiTA  VKELR  TGAGMMDCK" alimmline="" alimodel="EiTAAMVKELR..TGAGMMDCK" aliniseqs="5" alintseq="" alippline="9*********555********9" alirfline="" alisindex="57516019" alisqacc="4pc3_C" alisqdesc="Elongation factor Ts" alisqfrom="2" alisqname="4pc3_C" alisqto="23" bias="1.14" bitscore="39.962963104248" cevalue="1.1e-13" iali="2" ienv="1" ievalue="7.0e-09" is_included="1" is_reported="1" jali="23" jenv="24" oasc="0.90" />
#>       <domains aliId="0.8" aliIdCount="4" aliL="282" aliM="23" aliN="5" aliSim="0.8" aliSimCount="4" aliaseq="AAMSK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="6" alihmmname="Query" alihmmto="10" alimline="AAM K" alimmline="" alimodel="AAMVK" aliniseqs="5" alintseq="" alippline="9**77" alirfline="" alisindex="57516019" alisqacc="4pc3_C" alisqdesc="Elongation factor Ts" alisqfrom="276" alisqname="4pc3_C" alisqto="280" bias="0.41" bitscore="-3.55217576026917" cevalue="5.6" iali="276" ienv="275" ievalue="340000" is_included="0" is_reported="0" jali="280" jenv="280" oasc="0.82" />
#>       <pdbs>4pc3_D</pdbs>
#>       <pdbs>4q7j_E</pdbs>
#>       <pdbs>4pc3_C</pdbs>
#>       <pdbs>4q7j_A</pdbs>
#>       <pdbs>4pc7_C</pdbs>
#>       <seqs an="" de="Elongation factor Ts" dn="4pc3_D" sp="Escherichia coli K-12" />
#>       <seqs an="" de="Elongation factor Ts" dn="4pc7_C" sp="Escherichia coli K-12" />
#>       <seqs an="" de="Elongation factor Ts" dn="4q7j_A" sp="Escherichia coli K-12" />
#>       <seqs an="" de="Elongation factor Ts" dn="4q7j_E" sp="Escherichia coli K-12" />
#>     </hits>
#>     <hits name="1efu_B" acc="1efu_B" arch="PF00889.22" archScore="2" archindex="158710124293725 562" bias="1.1" desc="ELONGATION FACTOR TS" evalue="7.0e-09" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/1efu_B" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="1" ph="Proteobacteria" pvalue="-31.9955245624493" score="40.0" sindex="56418786" species="Escherichia coli" taxid="562" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.9" aliIdCount="18" aliL="282" aliM="23" aliN="22" aliSim="0.9" aliSimCount="18" aliaseq="EITASLVKELRerTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="3" alihmmname="Query" alihmmto="22" alimline="EiTA  VKELR  TGAGMMDCK" alimmline="" alimodel="EiTAAMVKELR..TGAGMMDCK" aliniseqs="6" alintseq="" alippline="9*********555********9" alirfline="" alisindex="56418786" alisqacc="1efu_B" alisqdesc="ELONGATION FACTOR TS" alisqfrom="2" alisqname="1efu_B" alisqto="23" bias="1.14" bitscore="39.962963104248" cevalue="1.1e-13" iali="2" ienv="1" ievalue="7.0e-09" is_included="1" is_reported="1" jali="23" jenv="24" oasc="0.90" />
#>       <domains aliId="0.8" aliIdCount="4" aliL="282" aliM="23" aliN="5" aliSim="0.8" aliSimCount="4" aliaseq="AAMSK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="6" alihmmname="Query" alihmmto="10" alimline="AAM K" alimmline="" alimodel="AAMVK" aliniseqs="6" alintseq="" alippline="9**77" alirfline="" alisindex="56418786" alisqacc="1efu_B" alisqdesc="ELONGATION FACTOR TS" alisqfrom="276" alisqname="1efu_B" alisqto="280" bias="0.41" bitscore="-3.55217576026917" cevalue="5.6" iali="276" ienv="275" ievalue="340000" is_included="0" is_reported="0" jali="280" jenv="280" oasc="0.82" />
#>       <pdbs>4pc1_D</pdbs>
#>       <pdbs>1efu_D</pdbs>
#>       <pdbs>4pc2_C</pdbs>
#>       <pdbs>4pc1_C</pdbs>
#>       <pdbs>1efu_B</pdbs>
#>       <pdbs>4pc2_D</pdbs>
#>       <seqs an="" de="ELONGATION FACTOR TS" dn="1efu_D" sp="Escherichia coli" />
#>       <seqs an="" de="Elongation factor Ts" dn="4pc1_C" sp="Escherichia coli" />
#>       <seqs an="" de="Elongation factor Ts" dn="4pc1_D" sp="Escherichia coli" />
#>       <seqs an="" de="Elongation factor Ts" dn="4pc2_C" sp="Escherichia coli" />
#>       <seqs an="" de="Elongation factor Ts" dn="4pc2_D" sp="Escherichia coli" />
#>     </hits>
#>     <hits name="3agp_A" acc="3agp_A" arch="PF00889.22 PF00009.30 PF03144.28 PF03143.20 PF03431.16" archScore="5" archindex="127492375402814 39803" bias="5.6" desc="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" evalue="1.1e-08" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/3agp_A" flags="3" kg="Viruses" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="2" ph="Lenarviricota" pvalue="-31.5560995902716" score="39.4" sindex="2984797" species="Escherichia virus Qbeta" taxid="39803" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.863636363636364" aliIdCount="19" aliL="1289" aliM="23" aliN="24" aliSim="0.863636363636364" aliSimCount="19" aliaseq="MAEITASLVKELRerTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="1" alihmmname="Query" alihmmto="22" alimline="m EiTA  VKELR  TGAGMMDCK" alimmline="" alimodel="mtEiTAAMVKELR..TGAGMMDCK" aliniseqs="11" alintseq="" alippline="99**********555********9" alirfline="" alisindex="2984797" alisqacc="3agp_A" alisqdesc="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" alisqfrom="1" alisqname="3agp_A" alisqto="24" bias="2.51" bitscore="38.9139556884766" cevalue="2.4e-13" iali="1" ienv="1" ievalue="1.5e-08" is_included="1" is_reported="1" jali="24" jenv="25" oasc="0.91" />
#>       <domains aliId="0.583333333333333" aliIdCount="7" aliL="1289" aliM="23" aliN="12" aliSim="0.666666666666667" aliSimCount="8" aliaseq="VKNMITGAAQMD" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="20" alimline="VK + TGA  MD" alimmline="" alimodel="VKELRTGAGMMD" aliniseqs="11" alintseq="" alippline="99999***9998" alirfline="" alisindex="2984797" alisqacc="3agp_A" alisqdesc="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" alisqfrom="373" alisqname="3agp_A" alisqto="384" bias="0.05" bitscore="0.0737475380301476" cevalue="0.41" iali="373" ienv="370" ievalue="25000" is_included="0" is_reported="1" jali="384" jenv="385" oasc="0.86" />
#>       <pdbs>3vnv_A</pdbs>
#>       <pdbs>3vnu_A</pdbs>
#>       <pdbs>3avu_A</pdbs>
#>       <pdbs>3agq_A</pdbs>
#>       <pdbs>3avt_A</pdbs>
#>       <pdbs>3avv_A</pdbs>
#>       <pdbs>3avx_A</pdbs>
#>       <pdbs>3avy_A</pdbs>
#>       <pdbs>3avw_A</pdbs>
#>       <pdbs>3agp_A</pdbs>
#>       <pdbs>4fwt_A</pdbs>
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu 1, LINKER, Q beta replicase" dn="3agq_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avt_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avu_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avv_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avw_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avx_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avy_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3vnu_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3vnv_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="4fwt_A" sp="Escherichia virus Qbeta" />
#>     </hits>
#>     <hits name="3agp_A" acc="3agp_A" arch="PF00889.22 PF00009.30 PF03144.28 PF03143.20 PF03431.16" archScore="5" archindex="127492375402814 83334" bias="5.6" desc="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" evalue="1.1e-08" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/3agp_A" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="2" ph="Proteobacteria" pvalue="-31.5560995902716" score="39.4" sindex="57523359" species="Escherichia coli O157:H7" taxid="83334" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.863636363636364" aliIdCount="19" aliL="1289" aliM="23" aliN="24" aliSim="0.863636363636364" aliSimCount="19" aliaseq="MAEITASLVKELRerTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="1" alihmmname="Query" alihmmto="22" alimline="m EiTA  VKELR  TGAGMMDCK" alimmline="" alimodel="mtEiTAAMVKELR..TGAGMMDCK" aliniseqs="11" alintseq="" alippline="99**********555********9" alirfline="" alisindex="57523359" alisqacc="3agp_A" alisqdesc="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" alisqfrom="1" alisqname="3agp_A" alisqto="24" bias="2.51" bitscore="38.9139556884766" cevalue="2.4e-13" iali="1" ienv="1" ievalue="1.5e-08" is_included="1" is_reported="1" jali="24" jenv="25" oasc="0.91" />
#>       <domains aliId="0.583333333333333" aliIdCount="7" aliL="1289" aliM="23" aliN="12" aliSim="0.666666666666667" aliSimCount="8" aliaseq="VKNMITGAAQMD" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="20" alimline="VK + TGA  MD" alimmline="" alimodel="VKELRTGAGMMD" aliniseqs="11" alintseq="" alippline="99999***9998" alirfline="" alisindex="57523359" alisqacc="3agp_A" alisqdesc="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" alisqfrom="373" alisqname="3agp_A" alisqto="384" bias="0.05" bitscore="0.0737475380301476" cevalue="0.41" iali="373" ienv="370" ievalue="25000" is_included="0" is_reported="1" jali="384" jenv="385" oasc="0.86" />
#>       <pdbs>3vnv_A</pdbs>
#>       <pdbs>3vnu_A</pdbs>
#>       <pdbs>3avu_A</pdbs>
#>       <pdbs>3agq_A</pdbs>
#>       <pdbs>3avt_A</pdbs>
#>       <pdbs>3avv_A</pdbs>
#>       <pdbs>3avx_A</pdbs>
#>       <pdbs>3avy_A</pdbs>
#>       <pdbs>3avw_A</pdbs>
#>       <pdbs>3agp_A</pdbs>
#>       <pdbs>4fwt_A</pdbs>
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu 1, LINKER, Q beta replicase" dn="3agq_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avt_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avu_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avv_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avw_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avx_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avy_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3vnu_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3vnv_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="4fwt_A" sp="Escherichia coli O157:H7" />
#>     </hits>
#>     <hits name="1aip_C" acc="1aip_C" arch="PF00889.22" archScore="0" archindex="158710124293725 274" bias="0.4" desc="ELONGATION FACTOR TS" evalue="0.25" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/1aip_C" flags="2" kg="Bacteria" ndom="2" nincluded="0" niseqs="16" nregions="2" nreported="1" ph="Deinococcus-Thermus" pvalue="-14.6003256082915" score="16.0" sindex="153001630" species="Thermus thermophilus" taxid="274" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.785714285714286" aliIdCount="11" aliL="196" aliM="23" aliN="16" aliSim="0.857142857142857" aliSimCount="12" aliaseq="IKKLReaTGAGMMDVK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="22" alimline="+K LR  TGAGMMD K" alimmline="" alimodel="VKELR..TGAGMMDCK" aliniseqs="4" alintseq="" alippline="6888888*******98" alirfline="" alisindex="153001630" alisqacc="1aip_C" alisqdesc="ELONGATION FACTOR TS" alisqfrom="7" alisqname="1aip_C" alisqto="22" bias="0.35" bitscore="15.9513692855835" cevalue="4.1e-06" iali="7" ienv="5" ievalue="0.25" is_included="0" is_reported="1" jali="22" jenv="23" oasc="0.86" />
#>       <domains aliId="1" aliIdCount="4" aliL="196" aliM="23" aliN="4" aliSim="1" aliSimCount="4" aliaseq="VKEL" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="12" alimline="VKEL" alimmline="" alimodel="VKEL" aliniseqs="4" alintseq="" alippline="9**9" alirfline="" alisindex="153001630" alisqacc="1aip_C" alisqdesc="ELONGATION FACTOR TS" alisqfrom="169" alisqname="1aip_C" alisqto="172" bias="0.23" bitscore="-3.05213689804077" cevalue="3.9" iali="169" ienv="169" ievalue="240000" is_included="0" is_reported="0" jali="172" jenv="172" oasc="0.93" />
#>       <pdbs>1aip_C</pdbs>
#>       <pdbs>1aip_D</pdbs>
#>       <pdbs>1aip_H</pdbs>
#>       <pdbs>1aip_G</pdbs>
#>       <seqs an="" de="ELONGATION FACTOR TS" dn="1aip_D" sp="Thermus thermophilus" />
#>       <seqs an="" de="ELONGATION FACTOR TS" dn="1aip_G" sp="Thermus thermophilus" />
#>       <seqs an="" de="ELONGATION FACTOR TS" dn="1aip_H" sp="Thermus thermophilus" />
#>     </hits>
#>     <hits name="5w70_A" acc="5w70_A" arch="PF01041.20" archScore="0" archindex="207269702100274 80859" bias="0.1" desc="L-glutamine:2-deoxy-scyllo-inosose aminotransferase" evalue="0.47" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/5w70_A" flags="2" kg="Bacteria" ndom="1" nincluded="0" niseqs="16" nregions="1" nreported="1" ph="Actinobacteria" pvalue="-13.9680672002281" score="15.1" sindex="193492204" species="Streptomyces ribosidificus" taxid="80859" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.578947368421053" aliIdCount="11" aliL="445" aliM="23" aliN="19" aliSim="0.631578947368421" aliSimCount="12" aliaseq="MGDIAAAFAKVLRHGAGLM" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="1" alihmmname="Query" alihmmto="19" alimline="m +i AA  K LR GAG M" alimmline="" alimodel="mtEiTAAMVKELRTGAGMM" aliniseqs="2" alintseq="" alippline="789****************" alirfline="" alisindex="193492204" alisqacc="5w70_A" alisqdesc="L-glutamine:2-deoxy-scyllo-inosose aminotransferase" alisqfrom="426" alisqname="5w70_A" alisqto="444" bias="0.09" bitscore="13.8441429138184" cevalue="1.9e-05" iali="426" ienv="426" ievalue="1.1" is_included="0" is_reported="1" jali="444" jenv="445" oasc="0.95" />
#>       <pdbs>5w70_B</pdbs>
#>       <pdbs>5w70_A</pdbs>
#>       <seqs an="" de="L-glutamine:2-deoxy-scyllo-inosose aminotransferase" dn="5w70_B" sp="Streptomyces ribosidificus" />
#>     </hits>
#>     <stats Z="546676" Z_setby="0" domZ="9" domZ_setby="0" elapsed="0.05" n_past_bias="2380" n_past_fwd="15" n_past_msv="2862" n_past_vit="102" nhits="9" nincluded="7" nmodels="1" nreported="9" nseqs="546676" page="1" sys="0" total="1" unpacked="9" user="0" />
#>   </data>
#> </opt>
jackhmmer_tbl_using_aln
#> An object of class 'AnnotatedDataFrame'
#>   HMMERqueryNames: 1
#>   varLabels: dbs uuid ... phylip.url (17 total)
#>   varMetadata: labelDescription algorithm
Biobase::varMetadata(jackhmmer_tbl_using_aln)
#>                                                                                  labelDescription
#> dbs                                                  Names of target databases used in the search
#> uuid                                                                    The unique job identifier
#> score.url                                                            URL with HMMER results score
#> stats                                                                              The stats hash
#> hits                                                                     Array of sequence hashes
#> domains                                                          Array of sequence domains hashes
#> text.url                              A plain text file containing the hit alignments and scores.
#> tab.url                  A tab delimited text file containing the hit information. No alignments.
#> xml.url                                    An XML file formatted for machine parsing of the data.
#> json.url                             All the results information encoded as a single JSON string.
#> fasta.url                 Download the significant hits from your search as a gzipped FASTA file.
#> fullfasta.url                                                                   Full length FASTA
#> alignedfasta.url A gzipped file containing the full length sequences for significant search hits.
#> stockholm.url          A gzipped file containing aligned significant search hits in FASTA format.
#> clustalw.url                Download an alignment of significant hits as a gzipped ClustalW file.
#> psiblast.url                Download an alignment of significant hits as a gzipped psiblast file.
#> phylip.url                    Download an alignment of significant hits as a gzipped phylip file.
#>                  algorithm
#> dbs              jackhmmer
#> uuid             jackhmmer
#> score.url        jackhmmer
#> stats            jackhmmer
#> hits             jackhmmer
#> domains          jackhmmer
#> text.url         jackhmmer
#> tab.url          jackhmmer
#> xml.url          jackhmmer
#> json.url         jackhmmer
#> fasta.url        jackhmmer
#> fullfasta.url    jackhmmer
#> alignedfasta.url jackhmmer
#> stockholm.url    jackhmmer
#> clustalw.url     jackhmmer
#> psiblast.url     jackhmmer
#> phylip.url       jackhmmer