Perform a hmmscan search of a protein sequence against a profile-HMM database.

Perform a hmmscan search of a protein sequence against a profile-HMM database.

search_hmmscan(seq, hmmdb = "pfam", timeout = 180, verbose = FALSE)

Arguments

seq

A character vector containing the sequences of the query or any other object that can be converted to that.

hmmdb

A character vector containing the target databases. Frequently used databases are pfam, tigrfam gene3d, superfamily, pirsf and treefam, but a complete and updated list is available at https://www.ebi.ac.uk/Tools/hmmer/.

timeout

Set maximum request time in seconds.

verbose

A logical, if TRUE details of the download process is printed.

Value

An Data Frame containing the results from HMMER.

Examples

search_hmmscan(
    seq = "SWYHGPVSRNAAEYLLSSGINGSFLVRESESSPGQRSISLRYEGRVYHYRINTASDGKLYVSS",
    hmmdb = "pfam",
    verbose = FALSE
)
#> # A tibble: 1 × 36
#>   algori…¹ uuid  stats…² stats…³ stats…⁴ stats.Z stats…⁵ stats…⁶ stats…⁷ stats…⁸
#>   <chr>    <chr>   <dbl>   <int> <chr>     <dbl>   <int>   <int>   <int>   <int>
#> 1 hmmscan  2361…       1       1 0.06      19632       0     316       1       1
#> # … with 26 more variables: stats.user <dbl>, stats.domZ_setby <int>,
#> #   stats.n_past_bias <int>, stats.sys <dbl>, stats.n_past_fwd <int>,
#> #   stats.total <dbl>, stats.nmodels <int>, stats.nincluded <int>,
#> #   stats.n_past_vit <int>, stats.nreported <int>, stats.domZ <dbl>,
#> #   hits.flags <int>, hits.nregions <int>, hits.ndom <int>, hits.name <chr>,
#> #   hits.score <dbl>, hits.bias <chr>, hits.taxid <chr>, hits.acc <chr>,
#> #   hits.domains <list>, hits.nincluded <int>, hits.evalue <dbl>, …