Perform a hmmsearch search of a protein alignment against a protein sequence database.

Perform a hmmsearch search of a protein alignment against a protein sequence database.

search_hmmsearch(
  alns,
  dbs = "swissprot",
  fullseqfasta = TRUE,
  verbose = TRUE,
  timeout = 90,
  alignment = FALSE
)

Arguments

alns

A list of Biostrings::AAMultipleAlignment.

dbs

A character vector containing the target databases. Frequently used databases are swissprot, uniprotrefprot, uniprotkb, ensembl, pdb and alphafold, but a complete and updated list is available at https://www.ebi.ac.uk/Tools/hmmer/.

fullseqfasta

A logical, if TRUE full sequence fasta will be downloaded as a AAStringSet. Package Biostrings must be installed in that case.

verbose

A logical, if TRUE details of the download process is printed.

timeout

An integer specifying the number of seconds to wait for the reply before a time out occurs.

alignment

A logical, if TRUE sequence alignment will be downloaded as a AAMultipleAlignment. Package Biostrings must be installed in that case.#'

Value

A nested DataFrame with columns alns, dbs, url(HMMER temporary url), hits, stats, domains and, if selected, fullseq.fasta. and alignment.

Examples

# \donttest{
aln <- c(
    "MTEITAAMVKELR--TGAGMMDCKN",
    "------AMVKELRESTGAGMMDCKN"
) %>%
    Biostrings::AAMultipleAlignment()
hmmsearch_tbl <- search_hmmsearch(
    alns = aln,
    dbs = "pdb",
    verbose = FALSE,
    timeout = 90
)
#> Content from HMMER server:
#> <opt>
#>   <data name="results" algo="hmmsearch" uuid="3CF9980C-B838-11EC-8622-C365F75AEC3D">
#>     <_internal highevalue="1.7e-11" lowevalue="0.00024" />
#>     <hits name="4pc3_C" acc="4pc3_C" arch="PF00889.22" archScore="2" archindex="158710124293725 83333" bias="1.6" desc="Elongation factor Ts" evalue="1.7e-11" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/4pc3_C" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="1" ph="Proteobacteria" pvalue="-38.0097949688518" score="48.6" sindex="57516019" species="Escherichia coli K-12" taxid="83333" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.863636363636364" aliIdCount="19" aliL="282" aliM="25" aliN="22" aliSim="0.863636363636364" aliSimCount="19" aliaseq="EITASLVKELRERTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="3" alihmmname="Query" alihmmto="24" alimline="EitA  VKELRE TGAGMMDCK" alimmline="" alimodel="EitAAMVKELREsTGAGMMDCK" aliniseqs="5" alintseq="" alippline="9********************9" alirfline="" alisindex="57516019" alisqacc="4pc3_C" alisqdesc="Elongation factor Ts" alisqfrom="2" alisqname="4pc3_C" alisqto="23" bias="1.64" bitscore="48.611629486084" cevalue="3.4e-16" iali="2" ienv="1" ievalue="1.7e-11" is_included="1" is_reported="1" jali="23" jenv="24" oasc="0.97" />
#>       <domains aliId="0.8" aliIdCount="4" aliL="282" aliM="25" aliN="5" aliSim="0.8" aliSimCount="4" aliaseq="AAMSK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="6" alihmmname="Query" alihmmto="10" alimline="AAM K" alimmline="" alimodel="AAMVK" aliniseqs="5" alintseq="" alippline="9**77" alirfline="" alisindex="57516019" alisqacc="4pc3_C" alisqdesc="Elongation factor Ts" alisqfrom="276" alisqname="4pc3_C" alisqto="280" bias="0.38" bitscore="-3.95390582084656" cevalue="10" iali="276" ienv="275" ievalue="500000" is_included="0" is_reported="0" jali="280" jenv="280" oasc="0.80" />
#>       <pdbs>4pc3_D</pdbs>
#>       <pdbs>4q7j_E</pdbs>
#>       <pdbs>4pc3_C</pdbs>
#>       <pdbs>4q7j_A</pdbs>
#>       <pdbs>4pc7_C</pdbs>
#>       <seqs an="" de="Elongation factor Ts" dn="4pc3_D" sp="Escherichia coli K-12" />
#>       <seqs an="" de="Elongation factor Ts" dn="4pc7_C" sp="Escherichia coli K-12" />
#>       <seqs an="" de="Elongation factor Ts" dn="4q7j_A" sp="Escherichia coli K-12" />
#>       <seqs an="" de="Elongation factor Ts" dn="4q7j_E" sp="Escherichia coli K-12" />
#>     </hits>
#>     <hits name="1efu_B" acc="1efu_B" arch="PF00889.22" archScore="2" archindex="158710124293725 562" bias="1.6" desc="ELONGATION FACTOR TS" evalue="1.7e-11" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/1efu_B" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="1" ph="Proteobacteria" pvalue="-38.0097949688518" score="48.6" sindex="56418786" species="Escherichia coli" taxid="562" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.863636363636364" aliIdCount="19" aliL="282" aliM="25" aliN="22" aliSim="0.863636363636364" aliSimCount="19" aliaseq="EITASLVKELRERTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="3" alihmmname="Query" alihmmto="24" alimline="EitA  VKELRE TGAGMMDCK" alimmline="" alimodel="EitAAMVKELREsTGAGMMDCK" aliniseqs="6" alintseq="" alippline="9********************9" alirfline="" alisindex="56418786" alisqacc="1efu_B" alisqdesc="ELONGATION FACTOR TS" alisqfrom="2" alisqname="1efu_B" alisqto="23" bias="1.64" bitscore="48.611629486084" cevalue="3.4e-16" iali="2" ienv="1" ievalue="1.7e-11" is_included="1" is_reported="1" jali="23" jenv="24" oasc="0.97" />
#>       <domains aliId="0.8" aliIdCount="4" aliL="282" aliM="25" aliN="5" aliSim="0.8" aliSimCount="4" aliaseq="AAMSK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="6" alihmmname="Query" alihmmto="10" alimline="AAM K" alimmline="" alimodel="AAMVK" aliniseqs="6" alintseq="" alippline="9**77" alirfline="" alisindex="56418786" alisqacc="1efu_B" alisqdesc="ELONGATION FACTOR TS" alisqfrom="276" alisqname="1efu_B" alisqto="280" bias="0.38" bitscore="-3.95390582084656" cevalue="10" iali="276" ienv="275" ievalue="500000" is_included="0" is_reported="0" jali="280" jenv="280" oasc="0.80" />
#>       <pdbs>4pc1_D</pdbs>
#>       <pdbs>1efu_D</pdbs>
#>       <pdbs>4pc2_C</pdbs>
#>       <pdbs>4pc1_C</pdbs>
#>       <pdbs>1efu_B</pdbs>
#>       <pdbs>4pc2_D</pdbs>
#>       <seqs an="" de="ELONGATION FACTOR TS" dn="1efu_D" sp="Escherichia coli" />
#>       <seqs an="" de="Elongation factor Ts" dn="4pc1_C" sp="Escherichia coli" />
#>       <seqs an="" de="Elongation factor Ts" dn="4pc1_D" sp="Escherichia coli" />
#>       <seqs an="" de="Elongation factor Ts" dn="4pc2_C" sp="Escherichia coli" />
#>       <seqs an="" de="Elongation factor Ts" dn="4pc2_D" sp="Escherichia coli" />
#>     </hits>
#>     <hits name="4pc6_C" acc="4pc6_C" arch="PF00889.22" archScore="2" archindex="158710124293725 536056" bias="1.6" desc="Elongation factor Ts" evalue="1.7e-11" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/4pc6_C" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="1" ph="Proteobacteria" pvalue="-38.0097949688518" score="48.6" sindex="55721010" species="Escherichia coli (strain ATCC 33849 / DSM 4235 / NCIMB 12045 / K12 / DH1)" taxid="536056" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.863636363636364" aliIdCount="19" aliL="282" aliM="25" aliN="22" aliSim="0.863636363636364" aliSimCount="19" aliaseq="EITASLVKELRERTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="3" alihmmname="Query" alihmmto="24" alimline="EitA  VKELRE TGAGMMDCK" alimmline="" alimodel="EitAAMVKELREsTGAGMMDCK" aliniseqs="2" alintseq="" alippline="9********************9" alirfline="" alisindex="55721010" alisqacc="4pc6_C" alisqdesc="Elongation factor Ts" alisqfrom="2" alisqname="4pc6_C" alisqto="23" bias="1.64" bitscore="48.611629486084" cevalue="3.4e-16" iali="2" ienv="1" ievalue="1.7e-11" is_included="1" is_reported="1" jali="23" jenv="24" oasc="0.97" />
#>       <domains aliId="0.8" aliIdCount="4" aliL="282" aliM="25" aliN="5" aliSim="0.8" aliSimCount="4" aliaseq="AAMSK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="6" alihmmname="Query" alihmmto="10" alimline="AAM K" alimmline="" alimodel="AAMVK" aliniseqs="2" alintseq="" alippline="9**77" alirfline="" alisindex="55721010" alisqacc="4pc6_C" alisqdesc="Elongation factor Ts" alisqfrom="276" alisqname="4pc6_C" alisqto="280" bias="0.38" bitscore="-3.95390582084656" cevalue="10" iali="276" ienv="275" ievalue="500000" is_included="0" is_reported="0" jali="280" jenv="280" oasc="0.80" />
#>       <pdbs>4pc6_D</pdbs>
#>       <pdbs>4pc6_C</pdbs>
#>       <seqs an="" de="Elongation factor Ts" dn="4pc6_D" sp="Escherichia coli DH1" />
#>     </hits>
#>     <hits name="3mmp_A" acc="3mmp_A" arch="PF00889.22 PF00009.30 PF03144.28 PF03143.20" archScore="5" archindex="118498722050600 83333" bias="3.2" desc="Elongation factor Tu 2, Elongation factor Ts" evalue="2.1e-11" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/3mmp_A" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="1" ph="Proteobacteria" pvalue="-37.8169870091926" score="48.3" sindex="57514389" species="Escherichia coli K-12" taxid="83333" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.833333333333333" aliIdCount="20" aliL="678" aliM="25" aliN="24" aliSim="0.833333333333333" aliSimCount="20" aliaseq="MAEITASLVKELRERTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="1" alihmmname="Query" alihmmto="24" alimline="m EitA  VKELRE TGAGMMDCK" alimmline="" alimodel="mtEitAAMVKELREsTGAGMMDCK" aliniseqs="2" alintseq="" alippline="99*********************9" alirfline="" alisindex="57514389" alisqacc="3mmp_A" alisqdesc="Elongation factor Tu 2, Elongation factor Ts" alisqfrom="1" alisqname="3mmp_A" alisqto="24" bias="3.23" bitscore="48.3443603515625" cevalue="4.1e-16" iali="1" ienv="1" ievalue="2.1e-11" is_included="1" is_reported="1" jali="24" jenv="25" oasc="0.98" />
#>       <domains aliId="0.583333333333333" aliIdCount="7" aliL="678" aliM="25" aliN="14" aliSim="0.666666666666667" aliSimCount="8" aliaseq="VKNMI--TGAAQMD" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="22" alimline="VK +   TGA  MD" alimmline="" alimodel="VKELREsTGAGMMD" aliniseqs="2" alintseq="" alippline="77777..***9999" alirfline="" alisindex="57514389" alisqacc="3mmp_A" alisqdesc="Elongation factor Tu 2, Elongation factor Ts" alisqfrom="373" alisqname="3mmp_A" alisqto="384" bias="0.09" bitscore="-0.792493999004364" cevalue="1" iali="373" ienv="370" ievalue="51000" is_included="0" is_reported="0" jali="384" jenv="385" oasc="0.83" />
#>       <pdbs>3mmp_C</pdbs>
#>       <pdbs>3mmp_A</pdbs>
#>       <seqs an="" de="Elongation factor Tu 2, Elongation factor Ts" dn="3mmp_C" sp="Escherichia coli K-12" />
#>     </hits>
#>     <hits name="4r71_A" acc="4r71_A" arch="PF00889.22 PF00009.30 PF03144.28 PF03143.20" archScore="5" archindex="118498722050600 83333" bias="3.2" desc="Elongation factor Ts, Elongation factor Tu" evalue="2.1e-11" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/4r71_A" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="1" ph="Proteobacteria" pvalue="-37.7900869648658" score="48.3" sindex="57510549" species="Escherichia coli K-12" taxid="83333" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.833333333333333" aliIdCount="20" aliL="694" aliM="25" aliN="24" aliSim="0.833333333333333" aliSimCount="20" aliaseq="MAEITASLVKELRERTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="1" alihmmname="Query" alihmmto="24" alimline="m EitA  VKELRE TGAGMMDCK" alimmline="" alimodel="mtEitAAMVKELREsTGAGMMDCK" aliniseqs="2" alintseq="" alippline="99*********************9" alirfline="" alisindex="57510549" alisqacc="4r71_A" alisqdesc="Elongation factor Ts, Elongation factor Tu" alisqfrom="1" alisqname="4r71_A" alisqto="24" bias="3.23" bitscore="48.307071685791" cevalue="4.3e-16" iali="1" ienv="1" ievalue="2.1e-11" is_included="1" is_reported="1" jali="24" jenv="25" oasc="0.98" />
#>       <domains aliId="0.583333333333333" aliIdCount="7" aliL="694" aliM="25" aliN="14" aliSim="0.666666666666667" aliSimCount="8" aliaseq="VKNMI--TGAAQMD" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="22" alimline="VK +   TGA  MD" alimmline="" alimodel="VKELREsTGAGMMD" aliniseqs="2" alintseq="" alippline="77777..***9999" alirfline="" alisindex="57510549" alisqacc="4r71_A" alisqdesc="Elongation factor Ts, Elongation factor Tu" alisqfrom="373" alisqname="4r71_A" alisqto="384" bias="0.09" bitscore="-0.828099191188812" cevalue="1.1" iali="373" ienv="370" ievalue="52000" is_included="0" is_reported="0" jali="384" jenv="385" oasc="0.83" />
#>       <pdbs>4r71_C</pdbs>
#>       <pdbs>4r71_A</pdbs>
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu" dn="4r71_C" sp="Escherichia coli K-12" />
#>     </hits>
#>     <hits name="3agp_A" acc="3agp_A" arch="PF00889.22 PF00009.30 PF03144.28 PF03143.20 PF03431.16" archScore="5" archindex="127492375402814 39803" bias="3.2" desc="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" evalue="4.2e-11" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/3agp_A" flags="3" kg="Viruses" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="1" ph="Lenarviricota" pvalue="-37.0947393944964" score="47.3" sindex="2984797" species="Escherichia virus Qbeta" taxid="39803" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.833333333333333" aliIdCount="20" aliL="1289" aliM="25" aliN="24" aliSim="0.833333333333333" aliSimCount="20" aliaseq="MAEITASLVKELRERTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="1" alihmmname="Query" alihmmto="24" alimline="m EitA  VKELRE TGAGMMDCK" alimmline="" alimodel="mtEitAAMVKELREsTGAGMMDCK" aliniseqs="11" alintseq="" alippline="99*********************9" alirfline="" alisindex="2984797" alisqacc="3agp_A" alisqdesc="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" alisqfrom="1" alisqname="3agp_A" alisqto="24" bias="3.23" bitscore="47.3431854248047" cevalue="8.5e-16" iali="1" ienv="1" ievalue="4.2e-11" is_included="1" is_reported="1" jali="24" jenv="25" oasc="0.98" />
#>       <domains aliId="0.583333333333333" aliIdCount="7" aliL="1289" aliM="25" aliN="14" aliSim="0.666666666666667" aliSimCount="8" aliaseq="VKNMI--TGAAQMD" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="22" alimline="VK +   TGA  MD" alimmline="" alimodel="VKELREsTGAGMMD" aliniseqs="11" alintseq="" alippline="77777..***9999" alirfline="" alisindex="2984797" alisqacc="3agp_A" alisqdesc="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" alisqfrom="373" alisqname="3agp_A" alisqto="384" bias="0.09" bitscore="-1.75904405117035" cevalue="2.1" iali="373" ienv="370" ievalue="100000" is_included="0" is_reported="0" jali="384" jenv="385" oasc="0.83" />
#>       <pdbs>3vnv_A</pdbs>
#>       <pdbs>3vnu_A</pdbs>
#>       <pdbs>3avu_A</pdbs>
#>       <pdbs>3agq_A</pdbs>
#>       <pdbs>3avt_A</pdbs>
#>       <pdbs>3avv_A</pdbs>
#>       <pdbs>3avx_A</pdbs>
#>       <pdbs>3avy_A</pdbs>
#>       <pdbs>3avw_A</pdbs>
#>       <pdbs>3agp_A</pdbs>
#>       <pdbs>4fwt_A</pdbs>
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu 1, LINKER, Q beta replicase" dn="3agq_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avt_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avu_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avv_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avw_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avx_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avy_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3vnu_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3vnv_A" sp="Escherichia virus Qbeta" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="4fwt_A" sp="Escherichia virus Qbeta" />
#>     </hits>
#>     <hits name="3agp_A" acc="3agp_A" arch="PF00889.22 PF00009.30 PF03144.28 PF03143.20 PF03431.16" archScore="5" archindex="127492375402814 83334" bias="3.2" desc="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" evalue="4.2e-11" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/3agp_A" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="1" ph="Proteobacteria" pvalue="-37.0947393944964" score="47.3" sindex="57523359" species="Escherichia coli O157:H7" taxid="83334" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.833333333333333" aliIdCount="20" aliL="1289" aliM="25" aliN="24" aliSim="0.833333333333333" aliSimCount="20" aliaseq="MAEITASLVKELRERTGAGMMDCK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="1" alihmmname="Query" alihmmto="24" alimline="m EitA  VKELRE TGAGMMDCK" alimmline="" alimodel="mtEitAAMVKELREsTGAGMMDCK" aliniseqs="11" alintseq="" alippline="99*********************9" alirfline="" alisindex="57523359" alisqacc="3agp_A" alisqdesc="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" alisqfrom="1" alisqname="3agp_A" alisqto="24" bias="3.23" bitscore="47.3431854248047" cevalue="8.5e-16" iali="1" ienv="1" ievalue="4.2e-11" is_included="1" is_reported="1" jali="24" jenv="25" oasc="0.98" />
#>       <domains aliId="0.583333333333333" aliIdCount="7" aliL="1289" aliM="25" aliN="14" aliSim="0.666666666666667" aliSimCount="8" aliaseq="VKNMI--TGAAQMD" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="22" alimline="VK +   TGA  MD" alimmline="" alimodel="VKELREsTGAGMMD" aliniseqs="11" alintseq="" alippline="77777..***9999" alirfline="" alisindex="57523359" alisqacc="3agp_A" alisqdesc="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" alisqfrom="373" alisqname="3agp_A" alisqto="384" bias="0.09" bitscore="-1.75904405117035" cevalue="2.1" iali="373" ienv="370" ievalue="100000" is_included="0" is_reported="0" jali="384" jenv="385" oasc="0.83" />
#>       <pdbs>3vnv_A</pdbs>
#>       <pdbs>3vnu_A</pdbs>
#>       <pdbs>3avu_A</pdbs>
#>       <pdbs>3agq_A</pdbs>
#>       <pdbs>3avt_A</pdbs>
#>       <pdbs>3avv_A</pdbs>
#>       <pdbs>3avx_A</pdbs>
#>       <pdbs>3avy_A</pdbs>
#>       <pdbs>3avw_A</pdbs>
#>       <pdbs>3agp_A</pdbs>
#>       <pdbs>4fwt_A</pdbs>
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu 1, LINKER, Q beta replicase" dn="3agq_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avt_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avu_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avv_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avw_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avx_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3avy_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3vnu_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="3vnv_A" sp="Escherichia coli O157:H7" />
#>       <seqs an="" de="Elongation factor Ts, Elongation factor Tu, LINKER, Q beta replicase" dn="4fwt_A" sp="Escherichia coli O157:H7" />
#>     </hits>
#>     <hits name="1aip_C" acc="1aip_C" arch="PF00889.22" archScore="0" archindex="158710124293725 274" bias="1.1" desc="ELONGATION FACTOR TS" evalue="0.00024" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/1aip_C" flags="3" kg="Bacteria" ndom="2" nincluded="1" niseqs="16" nregions="2" nreported="1" ph="Deinococcus-Thermus" pvalue="-21.5467201678114" score="25.8" sindex="153001630" species="Thermus thermophilus" taxid="274" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.75" aliIdCount="12" aliL="196" aliM="25" aliN="16" aliSim="0.875" aliSimCount="14" aliaseq="IKKLREATGAGMMDVK" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="24" alimline="+K LRE+TGAGMMD K" alimmline="" alimodel="VKELREsTGAGMMDCK" aliniseqs="4" alintseq="" alippline="79************87" alirfline="" alisindex="153001630" alisqacc="1aip_C" alisqdesc="ELONGATION FACTOR TS" alisqfrom="7" alisqname="1aip_C" alisqto="22" bias="1.13" bitscore="25.7906227111816" cevalue="4.8e-09" iali="7" ienv="4" ievalue="0.00024" is_included="1" is_reported="1" jali="22" jenv="23" oasc="0.92" />
#>       <domains aliId="1" aliIdCount="4" aliL="196" aliM="25" aliN="4" aliSim="1" aliSimCount="4" aliaseq="VKEL" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="12" alimline="VKEL" alimmline="" alimodel="VKEL" aliniseqs="4" alintseq="" alippline="9**9" alirfline="" alisindex="153001630" alisqacc="1aip_C" alisqdesc="ELONGATION FACTOR TS" alisqfrom="169" alisqname="1aip_C" alisqto="172" bias="0.23" bitscore="-3.26252603530884" cevalue="6.1" iali="169" ienv="169" ievalue="300000" is_included="0" is_reported="0" jali="172" jenv="172" oasc="0.93" />
#>       <pdbs>1aip_C</pdbs>
#>       <pdbs>1aip_D</pdbs>
#>       <pdbs>1aip_H</pdbs>
#>       <pdbs>1aip_G</pdbs>
#>       <seqs an="" de="ELONGATION FACTOR TS" dn="1aip_D" sp="Thermus thermophilus" />
#>       <seqs an="" de="ELONGATION FACTOR TS" dn="1aip_G" sp="Thermus thermophilus" />
#>       <seqs an="" de="ELONGATION FACTOR TS" dn="1aip_H" sp="Thermus thermophilus" />
#>     </hits>
#>     <hits name="6w6p_Z" acc="6w6p_Z" arch="PF00831.26" archScore="1" archindex="98422932968362 474186" bias="0.3" desc="50S ribosomal protein L29" evalue="0.27" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/6w6p_Z" flags="2" kg="Bacteria" ndom="2" nincluded="0" niseqs="16" nregions="2" nreported="2" ph="Firmicutes" pvalue="-14.5156003412874" score="16.0" sindex="152225416" species="Enterococcus faecalis OG1RF" taxid="474186" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.642857142857143" aliIdCount="9" aliL="61" aliM="25" aliN="14" aliSim="0.642857142857143" aliSimCount="9" aliaseq="VKEIRELTTAEMLD" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="22" alimline="VKE RE T A M D" alimmline="" alimodel="VKELREsTGAGMMD" aliniseqs="2" alintseq="" alippline="9***********99" alirfline="" alisindex="152225416" alisqacc="6w6p_Z" alisqdesc="50S ribosomal protein L29" alisqfrom="2" alisqname="6w6p_Z" alisqto="15" bias="0.08" bitscore="11.4275417327881" cevalue="0.00015" iali="2" ienv="1" ievalue="7.6" is_included="0" is_reported="1" jali="15" jenv="16" oasc="0.96" />
#>       <domains aliId="0.555555555555556" aliIdCount="5" aliL="61" aliM="25" aliN="9" aliSim="0.666666666666667" aliSimCount="6" aliaseq="ARIKEVRQS" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="7" alihmmname="Query" alihmmto="15" alimline="A +KE R s" alimmline="" alimodel="AMVKELREs" aliniseqs="2" alintseq="" alippline="779****77" alirfline="" alisindex="152225416" alisqacc="6w6p_Z" alisqdesc="50S ribosomal protein L29" alisqfrom="40" alisqname="6w6p_Z" alisqto="48" bias="0.02" bitscore="2.26810646057129" cevalue="0.11" iali="40" ienv="37" ievalue="5600" is_included="0" is_reported="1" jali="48" jenv="49" oasc="0.82" />
#>       <pdbs>6w6p_Z</pdbs>
#>       <pdbs>6wu9_Z</pdbs>
#>       <seqs an="" de="50S ribosomal protein L29" dn="6wu9_Z" sp="Enterococcus faecalis OG1RF" />
#>     </hits>
#>     <hits name="6o8w_Z" acc="6o8w_Z" arch="PF00831.26" archScore="1" archindex="98422932968362 1351" bias="0.3" desc="50S ribosomal protein L29" evalue="0.27" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/6o8w_Z" flags="2" kg="Bacteria" ndom="2" nincluded="0" niseqs="16" nregions="2" nreported="2" ph="Firmicutes" pvalue="-14.5156003412874" score="16.0" sindex="151716744" species="Enterococcus faecalis" taxid="1351" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.642857142857143" aliIdCount="9" aliL="61" aliM="25" aliN="14" aliSim="0.642857142857143" aliSimCount="9" aliaseq="VKEIRELTTAEMLD" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="22" alimline="VKE RE T A M D" alimmline="" alimodel="VKELREsTGAGMMD" aliniseqs="5" alintseq="" alippline="9***********99" alirfline="" alisindex="151716744" alisqacc="6o8w_Z" alisqdesc="50S ribosomal protein L29" alisqfrom="2" alisqname="6o8w_Z" alisqto="15" bias="0.08" bitscore="11.4275417327881" cevalue="0.00015" iali="2" ienv="1" ievalue="7.6" is_included="0" is_reported="1" jali="15" jenv="16" oasc="0.96" />
#>       <domains aliId="0.555555555555556" aliIdCount="5" aliL="61" aliM="25" aliN="9" aliSim="0.666666666666667" aliSimCount="6" aliaseq="ARIKEVRQS" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="7" alihmmname="Query" alihmmto="15" alimline="A +KE R s" alimmline="" alimodel="AMVKELREs" aliniseqs="5" alintseq="" alippline="779****77" alirfline="" alisindex="151716744" alisqacc="6o8w_Z" alisqdesc="50S ribosomal protein L29" alisqfrom="40" alisqname="6o8w_Z" alisqto="48" bias="0.02" bitscore="2.26810646057129" cevalue="0.11" iali="40" ienv="37" ievalue="5600" is_included="0" is_reported="1" jali="48" jenv="49" oasc="0.82" />
#>       <pdbs>6o8x_Z</pdbs>
#>       <pdbs>6o8w_Z</pdbs>
#>       <pdbs>6o90_Z</pdbs>
#>       <pdbs>6o8z_Z</pdbs>
#>       <pdbs>6o8y_Z</pdbs>
#>       <seqs an="" de="50S ribosomal protein L29" dn="6o8x_Z" sp="Enterococcus faecalis" />
#>       <seqs an="" de="50S ribosomal protein L29" dn="6o8y_Z" sp="Enterococcus faecalis" />
#>       <seqs an="" de="50S ribosomal protein L29" dn="6o8z_Z" sp="Enterococcus faecalis" />
#>       <seqs an="" de="50S ribosomal protein L29" dn="6o90_Z" sp="Enterococcus faecalis" />
#>     </hits>
#>     <hits name="7nhk_1" acc="7nhk_1" arch="PF00831.26" archScore="1" archindex="98422932968362 1351" bias="0.8" desc="50S ribosomal protein L29" evalue="0.35" extlink="https://www.ebi.ac.uk/pdbe/entry/pdb/7nhk_1" flags="2" kg="Bacteria" ndom="2" nincluded="0" niseqs="16" nregions="2" nreported="2" ph="Firmicutes" pvalue="-14.2593357962439" score="15.7" sindex="151724911" species="Enterococcus faecalis" taxid="1351" taxlink="http://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=Info&amp;id=">
#>       <domains aliId="0.642857142857143" aliIdCount="9" aliL="62" aliM="25" aliN="14" aliSim="0.642857142857143" aliSimCount="9" aliaseq="VKEIRELTTAEMLD" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="9" alihmmname="Query" alihmmto="22" alimline="VKE RE T A M D" alimmline="" alimodel="VKELREsTGAGMMD" aliniseqs="1" alintseq="" alippline="9***********99" alirfline="" alisindex="151724911" alisqacc="7nhk_1" alisqdesc="50S ribosomal protein L29" alisqfrom="3" alisqname="7nhk_1" alisqto="16" bias="0.28" bitscore="11.4310731887817" cevalue="0.00015" iali="3" ienv="1" ievalue="7.6" is_included="0" is_reported="1" jali="16" jenv="17" oasc="0.95" />
#>       <domains aliId="0.555555555555556" aliIdCount="5" aliL="62" aliM="25" aliN="9" aliSim="0.666666666666667" aliSimCount="6" aliaseq="ARIKEVRQS" alicsline="" alihmmacc="" alihmmdesc="" alihmmfrom="7" alihmmname="Query" alihmmto="15" alimline="A +KE R s" alimmline="" alimodel="AMVKELREs" aliniseqs="1" alintseq="" alippline="779****77" alirfline="" alisindex="151724911" alisqacc="7nhk_1" alisqdesc="50S ribosomal protein L29" alisqfrom="41" alisqname="7nhk_1" alisqto="49" bias="0.02" bitscore="2.233314037323" cevalue="0.12" iali="41" ienv="38" ievalue="5800" is_included="0" is_reported="1" jali="49" jenv="50" oasc="0.82" />
#>       <pdbs>7nhk_1</pdbs>
#>       <pdbs>6o8x_Z</pdbs>
#>       <pdbs>6o8w_Z</pdbs>
#>       <pdbs>6w6p_Z</pdbs>
#>       <pdbs>6o90_Z</pdbs>
#>       <pdbs>6o8z_Z</pdbs>
#>       <pdbs>6o8y_Z</pdbs>
#>     </hits>
#>     <stats Z="546676" Z_setby="0" domZ="11" domZ_setby="0" elapsed="0.01" n_past_bias="2249" n_past_fwd="14" n_past_msv="2575" n_past_vit="211" nhits="11" nincluded="8" nmodels="1" nreported="11" nseqs="546676" page="1" sys="0" total="1" unpacked="11" user="0" />
#>   </data>
#> </opt>
# }