Perform a hmmsearch search of a protein sequence against a protein sequence database.

Perform a hmmsearch search of a protein sequence against a protein sequence database.

search_hmmsearch(aln, seqdb = "swissprot", timeout = 180, verbose = FALSE)

Arguments

aln

A Biostrings::AAMultipleAlignment or a list of Biostrings::AAMultipleAlignment.

seqdb

A character vector containing the target databases. Frequently used databases are swissprot, uniprotrefprot, uniprotkb, ensembl, pdb and alphafold, but a complete and updated list is available at https://www.ebi.ac.uk/Tools/hmmer/.

timeout

Set maximum request time in seconds.

verbose

A logical, if TRUE details of the download process is printed.

Value

An Data Frame containing the results from HMMER.

Examples

system.file("extdata", "alignment.fasta", package = "HMMERutils") %>%
    Biostrings::readAAMultipleAlignment() %>%
    search_hmmsearch(
        seqdb = "swissprot",
        timeout = 180,
        verbose = FALSE
    )
#> # A tibble: 110 × 47
#>    algor…¹ uuid  stats…² stats…³ stats…⁴ stats.Z stats…⁵ stats…⁶ stats…⁷ stats…⁸
#>    <chr>   <chr>   <dbl>   <int> <chr>     <dbl>   <int>   <int>   <int>   <int>
#>  1 hmmsea… 1860…       1     110 0.07     565928       0   19129     110  565928
#>  2 hmmsea… 1860…       1     110 0.07     565928       0   19129     110  565928
#>  3 hmmsea… 1860…       1     110 0.07     565928       0   19129     110  565928
#>  4 hmmsea… 1860…       1     110 0.07     565928       0   19129     110  565928
#>  5 hmmsea… 1860…       1     110 0.07     565928       0   19129     110  565928
#>  6 hmmsea… 1860…       1     110 0.07     565928       0   19129     110  565928
#>  7 hmmsea… 1860…       1     110 0.07     565928       0   19129     110  565928
#>  8 hmmsea… 1860…       1     110 0.07     565928       0   19129     110  565928
#>  9 hmmsea… 1860…       1     110 0.07     565928       0   19129     110  565928
#> 10 hmmsea… 1860…       1     110 0.07     565928       0   19129     110  565928
#> # … with 100 more rows, 37 more variables: stats.user <dbl>,
#> #   stats.domZ_setby <int>, stats.n_past_bias <int>, stats.sys <dbl>,
#> #   stats.n_past_fwd <int>, stats.total <dbl>, stats.nmodels <int>,
#> #   stats.nincluded <int>, stats.n_past_vit <int>, stats.nreported <int>,
#> #   stats.domZ <dbl>, hits.archScore <chr>, hits.ph <chr>, hits.arch <chr>,
#> #   hits.kg <chr>, hits.ndom <int>, hits.extlink <chr>, hits.acc2 <chr>,
#> #   hits.taxid <chr>, hits.acc <chr>, hits.taxlink <chr>, hits.desc <chr>, …