Perform a phmmer search of a protein sequence against a protein sequence database.

Perform a phmmer search of a protein sequence against a protein sequence database.

search_phmmer(
  seqs,
  seq_names = NULL,
  dbs = "swissprot",
  timeout = 180,
  verbose = FALSE
)

Arguments

seqs

A character vector containing the sequences of the query or any other object that can be converted to that.

seq_names

A character vector containing the names of the sequences of the query. By default seq_names = NULL.

dbs

A character vector containing the target databases. Frequently used databases are swissprot, uniprotrefprot, uniprotkb, ensembl, pdb and alphafold, but a complete and updated list is available at https://www.ebi.ac.uk/Tools/hmmer/.

timeout

Set maximum request time in seconds.

verbose

A logical, if TRUE details of the download process is printed.

Value

An AnnotatedDataFrame, consisting of 2 parts, a nested DataFrame with the search hashes, the download links of all available files and of the HMMER page where the results are hosted, and the metadata associated to this DataFrame. Although all available results are available here, we recommend using the extract_from_HMMER_data_tbl function to preprocess the data.

Examples

phmmer_tbl <- search_phmmer(
  seqs = "MTEITAAMVKELRTGAGMMDCKN",
  seq_names = "1efu_B",
  dbs = "pdb",
  verbose = FALSE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> OK (HTTP 200).
phmmer_tbl
#> An object of class 'AnnotatedDataFrame'
#>   HMMERqueryNames: 1
#>   varLabels: seq.name dbs ... phylip.url (18 total)
#>   varMetadata: labelDescription algorithm
Biobase::varMetadata(phmmer_tbl)
#>                                                                                  labelDescription
#> seq.name                                                    Names of sequences used in the search
#> dbs                                                  Names of target databases used in the search
#> uuid                                                                    The unique job identifier
#> score.url                                                            URL with HMMER results score
#> stats                                                                              The stats hash
#> hits                                                                     Array of sequence hashes
#> domains                                                          Array of sequence domains hashes
#> text.url                              A plain text file containing the hit alignments and scores.
#> tab.url                  A tab delimited text file containing the hit information. No alignments.
#> xml.url                                    An XML file formatted for machine parsing of the data.
#> json.url                             All the results information encoded as a single JSON string.
#> fasta.url                 Download the significant hits from your search as a gzipped FASTA file.
#> fullfasta.url                                                                   Full length FASTA
#> alignedfasta.url A gzipped file containing the full length sequences for significant search hits.
#> stockholm.url          A gzipped file containing aligned significant search hits in FASTA format.
#> clustalw.url                Download an alignment of significant hits as a gzipped ClustalW file.
#> psiblast.url                Download an alignment of significant hits as a gzipped psiblast file.
#> phylip.url                    Download an alignment of significant hits as a gzipped phylip file.
#>                  algorithm
#> seq.name            phmmer
#> dbs                 phmmer
#> uuid                phmmer
#> score.url           phmmer
#> stats               phmmer
#> hits                phmmer
#> domains             phmmer
#> text.url            phmmer
#> tab.url             phmmer
#> xml.url             phmmer
#> json.url            phmmer
#> fasta.url           phmmer
#> fullfasta.url       phmmer
#> alignedfasta.url    phmmer
#> stockholm.url       phmmer
#> clustalw.url        phmmer
#> psiblast.url        phmmer
#> phylip.url          phmmer